[exotica_levenberg_marquardt_solver] Split into source and header
